+++ js/renderer.ts
@@ -18,6 +18,30 @@ import {
 import {property} from '@wonderlandengine/api/decorators.js';
 import {mat4, vec3} from 'gl-matrix';
 import {ReactNode} from 'react';
+import type {
+    ValueType,
+    ValueTypeNoAuto,
+    Color,
+    YogaNodeProps,
+    TextProps,
+    RoundedRectangleProps,
+    MeshProps,
+    NineSliceProps,
+    ReactComp,
+} from './renderer-types.js';
+
+// Re-export types so other modules that import from './renderer.js' keep working
+export type {
+    ValueType,
+    ValueTypeNoAuto,
+    Color,
+    YogaNodeProps,
+    TextProps,
+    RoundedRectangleProps,
+    MeshProps,
+    NineSliceProps,
+    ReactComp,
+} from './renderer-types.js';
 
 import Reconciler, {Fiber, HostConfig} from 'react-reconciler';
 import {
@@ -78,128 +102,9 @@ import {
 import {computeMeshChildTransforms} from './renderer/mesh-transform-helpers.js';
 import {propsEqual} from './renderer/props-helpers.js';
 
-export type ValueType = number | 'auto' | `${number}%`;
-export type ValueTypeNoAuto = number | `${number}%`;
-export type Color = string | Float32Array | number;
-
 // TEXT_BASE_SIZE and DEFAULT_FONT_SIZE are provided by renderer-text-helpers
 
 let Y: Yoga | null = null;
-
-export interface YogaNodeProps {
-    height?: ValueType;
-    width?: ValueType;
-
-    alignContent?: Align;
-    alignItems?: Align;
-    alignSelf?: Align;
-    justifyContent?: Justify;
-
-    aspectRatio?: number;
-    display?: Display;
-
-    flex?: number;
-    flexDirection?: FlexDirection;
-    flexGrow?: number;
-    flexBasis?: ValueType;
-    flexShrink?: number;
-    flexWrap?: Wrap;
-
-    isReferenceBaseline?: boolean;
-
-    gap?: number;
-    columnGap?: number;
-    rowGap?: number;
-
-    border?: number;
-    borderTop?: number;
-    borderBottom?: number;
-    borderLeft?: number;
-    borderRight?: number;
-
-    margin?: ValueType;
-    marginTop?: ValueType;
-    marginBottom?: ValueType;
-    marginLeft?: ValueType;
-    marginRight?: ValueType;
-
-    maxHeight?: ValueTypeNoAuto;
-    maxWidth?: ValueTypeNoAuto;
-
-    minHeight?: ValueTypeNoAuto;
-    minWidth?: ValueTypeNoAuto;
-
-    overflow?: Overflow;
-
-    padding?: ValueTypeNoAuto;
-    paddingTop?: ValueTypeNoAuto;
-    paddingBottom?: ValueTypeNoAuto;
-    paddingLeft?: ValueTypeNoAuto;
-    paddingRight?: ValueTypeNoAuto;
-
-    /* Relative z value to add to the usual increment, to allow widgets to render on top/behind other widgets */
-    z?: number;
-
-    top?: ValueTypeNoAuto;
-    left?: ValueTypeNoAuto;
-    right?: ValueTypeNoAuto;
-    bottom?: ValueTypeNoAuto;
-    position?: PositionType;
-
-    onClick?: (e: {x: number; y: number; e: MouseEvent}) => void;
-    onUp?: (e: {x: number; y: number; e: PointerEvent}) => void;
-    onDown?: (e: {x: number; y: number; e: PointerEvent}) => void;
-    onMove?: (e: {x: number; y: number; e: PointerEvent}) => void;
-    onHover?: (e: {x: number; y: number; e: PointerEvent}) => void;
-    onUnhover?: (e: {x: number; y: number; e: PointerEvent}) => void;
-}
-
-/**
- * Properties for text components
- */
-export interface TextProps extends YogaNodeProps {
-    text?: string;
-    fontSize?: number;
-    material?: Material | null;
-    textAlign?: 'left' | 'center' | 'right';
-    textWrap?: 'none' | 'soft' | 'hard' | 'clip';
-}
-
-/**
- * Properties for roundedRectangle components
- */
-export interface RoundedRectangleProps extends YogaNodeProps {
-    /* Material for the rounded rectangle mesh */
-    material?: Material | null;
-    /* Material for the rounded rectangle border */
-    borderMaterial?: Material | null;
-    /* Rounding in pixel-like units */
-    rounding?: number;
-    /* Rounding resolution */
-    resolution?: number;
-    roundTopLeft?: boolean;
-    roundTopRight?: boolean;
-    roundBottomLeft?: boolean;
-    roundBottomRight?: boolean;
-}
-
-/**
- * Properties for mesh components
- */
-export interface MeshProps extends YogaNodeProps {
-    material?: Material | null;
-    mesh?: Mesh | null;
-}
-
-/**
- * Properties for nineSlice components
- */
-export interface NineSliceProps extends YogaNodeProps {
-    material?: Material | null;
-    texture?: Texture | null;
-    borderSize?: number;
-    borderTextureSize?: number;
-}
 
 export function destroyTreeForNode(child: NodeWrapper, ctx: Context) {
     const childCount = child.children.length ?? 0;
@@ -892,17 +797,7 @@ const HostConfig: HostConfig<
 };
 const reconcilerInstance = Reconciler(HostConfig);
 
-type ReactComp = Component & {
-    needsUpdate: boolean;
-    textMaterial: Material;
-    scaling: number[];
-    renderCallback: () => void;
-    callbacks: Record<string, any>;
-
-    setContext(c: Context): void;
-    updateLayout(): void;
-    render(): ReactNode;
-};
+// ReactComp type is exported from renderer-types.ts
 
 export enum UISpace {
     World = 0,
